+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- docdb engine version wasn't being pinned properly for the cluster or parameter group
+
+### Fixed
+
 - bug fix: `helpers.fromJson` couldn't handle relative paths. Replaced use of `require.resolve`
   with `path.resolve` and added a couple of tests.